chore!: move `default_is_opcode_supported` to `pwg.rs`
